fix(image): use bitcoin128.png from Bitcoin core's repo
This change moves the Bitcoin desktop/Applications image from the bitcoin.png in the repo to bitcoin128.png in Bitcoin Core's repo. Calling this a fix rather than feat since it patches a potential security attack vector where steganography could be used to hide data in the bitcoin image. It's safer for users to download the image from the Bitcoin Core repo.